requirements.txt
feedparser
schedule
ebooklib
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
google-generativeai
python-dotenv
openai-whisper